alterado de activitis para utilização de fragment,resolvido,bug de carregamento da última playlist acessada
