+++ app/src/main/java/com/example/soundplayer/presentation/adapter/SoundAdapter.kt
@@ -1,7 +1,6 @@
 package com.example.soundplayer.presentation.adapter
 
 import android.annotation.SuppressLint
-import android.content.Intent
 import android.graphics.Color
 import android.util.Log
 import android.view.LayoutInflater
@@ -16,12 +15,12 @@ import com.example.soundplayer.commons.extension.convertMilesSecondToMinSec
 import com.example.soundplayer.databinding.ItemSoundBinding
 import com.example.soundplayer.model.PlayList
 import com.example.soundplayer.model.Sound
-import com.example.soundplayer.presentation.SongPlayActivity
 import com.example.soundplayer.presentation.SoundViewModel
 
 class SoundAdapter(
     private val soundViewModel: SoundViewModel,
     val isUpdateList :(Boolean)->Unit,
+    val initNewFragment :()->Unit,
 ) :RecyclerView.Adapter<SoundAdapter.SoundViewHolder>() {
    private  var soundsPlayList: PlayList? = null
    private  var actualSound :Sound? = null
@@ -54,76 +53,106 @@ class SoundAdapter(
          binding.txvTitle.text =soudd.title
          binding.imageView.setImageURI(soudd.uriMediaAlbum)
 
-         if (soundSelecionados.isNotEmpty()){
+         configSelectionedItemApperence(soudd)
 
-             if (soundSelecionados.contains(soudd)){
-                binding.cardItemSound.setCardBackgroundColor(Color.GRAY)
-             }else{
-                 binding.cardItemSound.setCardBackgroundColor(ContextCompat.getColor(
-                     binding.root.context,R.color.colorSurfaceVariant
-                 ))
-             }
-         }else{
-             isUpdateList(false)
-             binding.cardItemSound.setCardBackgroundColor(ContextCompat.getColor(
-                 binding.root.context,R.color.colorSurfaceVariant
-             ))
-         }
-
-         if (soudd.uriMediaAlbum != null){
-             binding.imageView.setImageURI(soudd.uriMediaAlbum)
-             if (binding.imageView.drawable == null){
-                 binding.imageView.setImageResource(R.drawable.transferir)
-             }
-         }
-         if (actualSound!=null && actualSound?.title == soudd.title){
-             binding.txvTitle.setTextColor(ContextCompat.getColor(binding.root.context,R.color.red))
-             binding.txvTitle.isSelected = true
-             binding.lottieSoundAnimePlaying.isVisible =true
-         }else{
-             binding.txvTitle.setTextColor(ContextCompat.getColor(binding.root.context,R.color.my_primary))
-             binding.txvTitle.isSelected = false
-             binding.lottieSoundAnimePlaying.isVisible =false
-         }
+         configApperenceItem(soudd)
 
 
-         binding.idContraint.setOnClickListener {
+         clicItemEvent(position, soudd)
+
+         longPressEvent(soudd)
+     }
+
+        private fun configSelectionedItemApperence(soudd: Sound) {
+            if (soundSelecionados.isNotEmpty()) {
+
+                if (soundSelecionados.contains(soudd)) {
+                    binding.cardItemSound.setCardBackgroundColor(Color.GRAY)
+                } else {
+                    binding.cardItemSound.setCardBackgroundColor(
+                        ContextCompat.getColor(
+                            binding.root.context, R.color.colorSurfaceVariant
+                        )
+                    )
+                }
+            } else {
+                isUpdateList(false)
+                binding.cardItemSound.setCardBackgroundColor(
+                    ContextCompat.getColor(
+                        binding.root.context, R.color.colorSurfaceVariant
+                    )
+                )
+            }
+        }
+
+        private fun configApperenceItem(soudd: Sound) {
+            if (soudd.uriMediaAlbum != null) {
+                binding.imageView.setImageURI(soudd.uriMediaAlbum)
+                if (binding.imageView.drawable == null) {
+                    binding.imageView.setImageResource(R.drawable.transferir)
+                }
+            }
+            if (actualSound != null && actualSound?.title == soudd.title) {
+                binding.txvTitle.setTextColor(
+                    ContextCompat.getColor(
+                        binding.root.context,
+                        R.color.red
+                    )
+                )
+                binding.txvTitle.isSelected = true
+                binding.lottieSoundAnimePlaying.isVisible = true
+            } else {
+                binding.txvTitle.setTextColor(
+                    ContextCompat.getColor(
+                        binding.root.context,
+                        R.color.my_primary
+                    )
+                )
+                binding.txvTitle.isSelected = false
+                binding.lottieSoundAnimePlaying.isVisible = false
+            }
+        }
+
+        private fun clicItemEvent(position: Int, soudd: Sound) {
+            binding.idContraint.setOnClickListener {
                 if (soundSelecionados.isEmpty()) {
-                    if (soundsPlayList != null){
+                    if (soundsPlayList != null) {
                         soundsPlayList!!.currentMusicPosition = position
-                        val intent= Intent(binding.root.context, SongPlayActivity::class.java)
                         soundViewModel.getAllMusics(soundsPlayList!!)
-                        ContextCompat.startActivity(binding.root.context,intent,null)
-                    }else {
+                        initNewFragment()
+                    } else {
                         Toast.makeText(it.context, "PlayList null", Toast.LENGTH_SHORT).show()
                     }
-                }
-                 else{
-                    if (!soundSelecionados.contains(soudd)){
+                } else {
+                    if (!soundSelecionados.contains(soudd)) {
                         binding.cardItemSound.setCardBackgroundColor(Color.GRAY)
                         soundSelecionados.add(soudd)
-                    }else{
-                        binding.cardItemSound.setCardBackgroundColor(ContextCompat.getColor(
-                            binding.root.context,R.color.colorSurfaceVariant
-                        ))
+                    } else {
+                        binding.cardItemSound.setCardBackgroundColor(
+                            ContextCompat.getColor(
+                                binding.root.context, R.color.colorSurfaceVariant
+                            )
+                        )
                         soundSelecionados.remove(soudd)
-                        if (soundSelecionados.isEmpty())isUpdateList(false)
+                        if (soundSelecionados.isEmpty()) isUpdateList(false)
                     }
                     Log.i("INFO_", "bind: ${soundSelecionados.size}")
-                 }
+                }
 
-         }
+            }
+        }
 
-         binding.idContraint.setOnLongClickListener {view->
-                if (soundSelecionados.size == 0){
+        private fun longPressEvent(soudd: Sound) {
+            binding.idContraint.setOnLongClickListener { view ->
+                if (soundSelecionados.size == 0) {
                     isUpdateList(true)
                     soundSelecionados.add(soudd)
                     binding.cardItemSound.setCardBackgroundColor(Color.GRAY)
                     Log.i("INFO_", "bind: ${soundSelecionados.size}")
                 }
-             true
-         }
-     }
+                true
+            }
+        }
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SoundViewHolder {
