+++ app/src/main/res/layout/fragment_main.xml
@@ -5,23 +5,7 @@
     android:id="@+id/contstaintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MainActivity">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/myHostFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        app:defaultNavHost="true"
-        app:navGraph="@navigation/sound_navigation_graph"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.5" />
-
-    <!--
+    >
 
     <include
         android:id="@+id/includeSelectItem"
@@ -153,7 +137,9 @@
         android:src="@drawable/baseline_playlist_add_24"
         style="@style/ThemeOverlay.Material3.FloatingActionButton.Secondary"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        android:contentDescription="TODO" />
 
--->
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
