+++ app/src/main/res/values/strings.xml
@@ -16,4 +16,6 @@
     <string name="selecionar_items">Selecionar Items</string>
     <string name="escolha_o_playlist">Escolha o PlayList</string>
     <string name="total_de_m_sicas">Total de Músicas %1$s</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
